// frontend/src/app/layout/api/agravios_dtos/recurring.interface.ts
export interface Recurring {
    id?: number;
    uuid?: string;
    name?: string;
    basicCoste?: number;
    basicCosteHour?: number; 
    basicCosteYear?: number;  
    createUserId?: number;
    createAt?: string;
    updateUserId?: number | null;
    updateAt?: string | null;
    deleteUserId?: number | null;
    deleteAt?: string | null;
}